# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interesting Facts API
  description: API for managing interesting facts with Firebase Auth and Firestore
  version: 0.0.1
  contact:
    name: API Support

servers:
  - url: http://localhost:3001
    description: Local development

paths:
  /ping:
    get:
      summary: Health check
      operationId: ping
      tags: [Health]
      responses:
        '200':
          description: Server is alive
          content:
            text/plain:
              schema:
                type: string
                example: PONG

  /api/interesting-facts:
    get:
      summary: Get 3 random interesting facts
      operationId: getRandomFacts
      tags: [Interesting Facts]
      responses:
        '200':
          description: Array of 3 random facts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InterestingFact'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a new interesting fact
      operationId: createFact
      tags: [Interesting Facts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestingFactAdd'
      responses:
        '201':
          description: Fact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestingFact'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/interesting-facts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Fact ID

    get:
      summary: Get a fact by ID
      operationId: getFactById
      tags: [Interesting Facts]
      responses:
        '200':
          description: Single fact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestingFact'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      summary: Update a fact by ID
      operationId: updateFact
      tags: [Interesting Facts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestingFactAdd'
      responses:
        '200':
          description: Fact updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestingFact'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Delete a fact by ID
      operationId: deleteFact
      tags: [Interesting Facts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fact deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Interesting fact deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/docs:
    get:
      summary: Scalar API Reference UI
      operationId: apiDocs
      tags: [Documentation]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scalar API reference HTML
          content:
            text/html:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT token obtained from the React Native client

  schemas:
    InterestingFact:
      type: object
      properties:
        id:
          type: string
          example: abc123
        title:
          type: string
          example: Honey never spoils
        description:
          type: string
          example: Archaeologists found 3000-year-old honey in Egyptian tombs that was still edible
        studyLocation:
          type: string
          example: University of Copenhagen
        source:
          type: string
          format: uri
          example: https://example.com/research
      required: [id, title, description, studyLocation, source]

    InterestingFactAdd:
      type: object
      properties:
        title:
          type: string
          example: Honey never spoils
        description:
          type: string
          example: Archaeologists found 3000-year-old honey in Egyptian tombs that was still edible
        studyLocation:
          type: string
          example: University of Copenhagen
        source:
          type: string
          format: uri
          example: https://example.com/research
      required: [title, description, studyLocation, source]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
        code:
          type: string
          example: NOT_FOUND
        details:
          type: object
          description: Additional error context (validation errors only)
      required: [error, code]

  responses:
    UnauthorizedError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Authorization header with Bearer token is required
            code: UNAUTHORIZED

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Interesting fact with id abc123 not found
            code: NOT_FOUND

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: A fact with the same title or description already exists
            code: VALIDATION_ERROR
            details:
              title: Duplicate title or description

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
            code: INTERNAL_ERROR